# openapi.yaml
openapi: 3.0.3
info:
  title: Drivers Scout API
  version: 1.0.0
  description: |
    HTTP API for the Drivers Scout service, which archives iRacing member statistics
    as dated CSV snapshots and exposes deltas and growth leaders.
servers:
  - url: http://localhost:8000
    description: Local development server
components:
  securitySchemes:
    LicenseKeyAuth:
      type: apiKey
      in: header
      name: X-License-Key
      description: >-
        Active license key issued via the admin endpoints. Required for member and
        leader endpoints when `LICENSE_ADMIN_SECRET` is configured.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: License
      description: >-
        Alternative way to provide the license token via the `Authorization: Bearer`
        header.
    AdminSecretAuth:
      type: apiKey
      in: header
      name: X-Admin-Secret
      description: Secret token matching `LICENSE_ADMIN_SECRET`, required for admin license endpoints.
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
    LicenseStatus:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
          nullable: true
        active:
          type: boolean
        revoked_at:
          type: string
          format: date-time
          nullable: true
        valid:
          type: boolean
      required: [key, label, active, revoked_at, valid]
    LicenseRecord:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
          nullable: true
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
          nullable: true
      required: [key, label, active, created_at, revoked_at]
    SubscriptionCreate:
      type: object
      properties:
        webhook_url:
          type: string
          format: uri
        category:
          type: string
        min_irating:
          type: integer
          nullable: true
          minimum: 0
      required: [webhook_url, category]
    SubscriptionResponse:
      type: object
      properties:
        id:
          type: integer
        license_key:
          type: string
        category:
          type: string
        min_irating:
          type: integer
          nullable: true
        webhook_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
      required: [id, license_key, category, webhook_url, created_at]
    FetchCounts:
      type: object
      properties:
        counts:
          type: object
          additionalProperties:
            type: integer
          description: Number of rows stored per category.
      required: [counts]
    SyncMembersResult:
      type: object
      properties:
        upserted:
          type: integer
          description: Number of member rows inserted or updated.
      required: [upserted]
    MemberSnapshot:
      type: object
      properties:
        cust_id:
          type: integer
        category:
          type: string
        snapshot_date:
          type: string
          format: date
        fetched_at:
          type: string
          format: date-time
        driver:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        irating:
          type: integer
          nullable: true
        starts:
          type: integer
          nullable: true
        wins:
          type: integer
          nullable: true
      required: [cust_id, category, snapshot_date, fetched_at]
    MemberSnapshotEntry:
      type: object
      properties:
        cust_id:
          type: integer
        driver:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        irating:
          type: integer
          nullable: true
        starts:
          type: integer
          nullable: true
        wins:
          type: integer
          nullable: true
      required: [cust_id]
    MemberSnapshotsResponse:
      type: object
      properties:
        category:
          type: string
        snapshot_date:
          type: string
          format: date
        fetched_at:
          type: string
          format: date-time
        results:
          type: array
          items:
            $ref: '#/components/schemas/MemberSnapshotEntry'
        missing:
          type: array
          items:
            type: integer
      required: [category, snapshot_date, fetched_at, results, missing]
    MemberSummary:
      type: object
      properties:
        cust_id:
          type: integer
        display_name:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
      required: [cust_id]
    MemberSearchResponse:
      type: object
      properties:
        query:
          type: string
        limit:
          type: integer
        offset:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/MemberSummary'
      required: [query, limit, offset, results]
    DeltaResult:
      type: object
      properties:
        cust_id:
          type: integer
        category:
          type: string
        start_date_used:
          type: string
          format: date
        end_date_used:
          type: string
          format: date
        start_value:
          type: integer
        end_value:
          type: integer
        delta:
          type: integer
        percent_change:
          type: number
          format: float
          nullable: true
      required: [cust_id, category, start_date_used, end_date_used, start_value, end_value, delta]
    TopGrowerEntry:
      type: object
      properties:
        cust_id:
          type: integer
        category:
          type: string
        end_value:
          type: integer
        delta:
          type: integer
        percent_change:
          type: number
          format: float
          nullable: true
        driver:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        starts:
          type: integer
          nullable: true
        wins:
          type: integer
          nullable: true
      required: [cust_id, category, end_value, delta]
    TopGrowersResponse:
      type: object
      properties:
        category:
          type: string
        min_current_irating:
          type: integer
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/TopGrowerEntry'
        snapshot_age_days:
          type: integer
          nullable: true
        start_date_used:
          type: string
          format: date
          nullable: true
        end_date_used:
          type: string
          format: date
          nullable: true
      required: [category, results]
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Service is running.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /licenses/{license_key}/status:
    get:
      summary: Check license status
      parameters:
        - in: path
          name: license_key
          required: true
          schema:
            type: string
      responses:
        '200':
          description: License validity information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LicenseStatus'
  /admin/run-fetch:
    post:
      summary: Trigger an immediate snapshot fetch
      description: Fetches the configured category snapshots for today and stores them locally.
      security:
        - AdminSecretAuth: []
      parameters:
        - in: query
          name: category
          required: false
          schema:
            type: string
          description: Category to fetch. Defaults to configured categories when omitted.
      responses:
        '200':
          description: Fetch completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FetchCounts'
  /admin/sync-members:
    post:
      summary: Sync members from latest snapshots
      description: Creates or updates Member records using the latest snapshot for each category.
      security:
        - AdminSecretAuth: []
      responses:
        '200':
          description: Member sync completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncMembersResult'
  /admin/discord-subscriptions/run:
    post:
      summary: Deliver Discord webhook notifications
      security:
        - AdminSecretAuth: []
      responses:
        '200':
          description: Discord subscriptions processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /admin/licenses:
    post:
      summary: Issue a new license
      security:
        - AdminSecretAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                  nullable: true
      responses:
        '200':
          description: License created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LicenseRecord'
        '401':
          description: Admin secret missing or invalid.
    get:
      summary: List licenses
      security:
        - AdminSecretAuth: []
      parameters:
        - in: query
          name: include_inactive
          schema:
            type: boolean
          description: Include revoked licenses when true.
      responses:
        '200':
          description: Licenses matching the filter.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LicenseRecord'
        '401':
          description: Admin secret missing or invalid.
  /admin/licenses/{license_key}/revoke:
    post:
      summary: Revoke a license key
      security:
        - AdminSecretAuth: []
      parameters:
        - in: path
          name: license_key
          required: true
          schema:
            type: string
      responses:
        '200':
          description: License revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LicenseRecord'
        '401':
          description: Admin secret missing or invalid.
        '404':
          description: License not found.
  /admin/licenses/{license_key}/activate:
    post:
      summary: Activate a license key
      security:
        - AdminSecretAuth: []
      parameters:
        - in: path
          name: license_key
          required: true
          schema:
            type: string
      responses:
        '200':
          description: License activated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LicenseRecord'
        '401':
          description: Admin secret missing or invalid.
        '404':
          description: License not found.
  /subscriptions:
    get:
      summary: List webhook subscriptions
      security:
        - LicenseKeyAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Subscriptions listed.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionResponse'
        '401':
          description: License missing or invalid.
    post:
      summary: Create or update a webhook subscription
      security:
        - LicenseKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionCreate'
      responses:
        '200':
          description: Subscription updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '201':
          description: Subscription created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '401':
          description: License missing or invalid.
        '422':
          description: Unsupported category.
  /subscriptions/{subscription_id}:
    delete:
      summary: Delete a webhook subscription
      security:
        - LicenseKeyAuth: []
        - BearerAuth: []
      parameters:
        - in: path
          name: subscription_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Subscription removed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '401':
          description: License missing or invalid.
        '404':
          description: Subscription not found.
  /members/{cust_id}/latest:
    get:
      summary: Latest snapshot for a member
      security:
        - LicenseKeyAuth: []
        - BearerAuth: []
      parameters:
        - in: path
          name: cust_id
          required: true
          schema:
            type: integer
        - in: query
          name: category
          schema:
            type: string
            default: sports_car
          description: Category name as configured in the service.
      responses:
        '200':
          description: Latest available snapshot for the member.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberSnapshot'
        '401':
          description: License missing or invalid.
        '400':
          description: Unsupported category.
        '404':
          description: No snapshot found.
  /members/latest:
    get:
      summary: Latest snapshot for multiple members
      security:
        - LicenseKeyAuth: []
        - BearerAuth: []
      parameters:
        - in: query
          name: cust_ids
          required: true
          schema:
            type: string
          description: Comma-separated list of member cust_ids (e.g., "111,222").
        - in: query
          name: category
          schema:
            type: string
            default: sports_car
          description: Category name as configured in the service.
      responses:
        '200':
          description: Latest available snapshot data for the requested members.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberSnapshotsResponse'
        '401':
          description: License missing or invalid.
        '400':
          description: Unsupported category or invalid cust_ids.
        '404':
          description: No snapshot found.
  /members/search:
    get:
      summary: Search members by display name
      security:
        - LicenseKeyAuth: []
        - BearerAuth: []
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
            minLength: 3
          description: Partial display name to search for.
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Maximum number of results to return.
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Number of results to skip for pagination.
      responses:
        '200':
          description: Matching members by display name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberSearchResponse'
        '401':
          description: License missing or invalid.
        '422':
          description: Query too short.
  /members/{cust_id}/delta:
    get:
      summary: iRating delta for a member
      security:
        - LicenseKeyAuth: []
        - BearerAuth: []
      parameters:
        - in: path
          name: cust_id
          required: true
          schema:
            type: integer
        - in: query
          name: category
          schema:
            type: string
            default: sports_car
        - in: query
          name: days
          schema:
            type: integer
            minimum: 1
          description: Convenience window size. When provided, start date is derived from end date minus days.
        - in: query
          name: start
          schema:
            type: string
            format: date
          description: Explicit start date (YYYY-MM-DD).
        - in: query
          name: end
          schema:
            type: string
            format: date
          description: Explicit end date (YYYY-MM-DD). Defaults to today.
      responses:
        '200':
          description: Delta between the two snapshots.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeltaResult'
        '401':
          description: License missing or invalid.
        '400':
          description: Unsupported category.
        '404':
          description: Insufficient data to compute delta.
  /leaders/growers:
    get:
      summary: Top iRating growers
      security:
        - LicenseKeyAuth: []
        - BearerAuth: []
      parameters:
        - in: query
          name: category
          schema:
            type: string
            default: sports_car
        - in: query
          name: days
          schema:
            type: integer
            minimum: 1
            default: 30
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: min_current_irating
          schema:
            type: integer
            minimum: 0
          description: Require the end snapshot to have at least this iRating.
      responses:
        '200':
          description: Leaders ordered by iRating gain.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopGrowersResponse'
        '401':
          description: License missing or invalid.
        '400':
          description: Unsupported category.